Implement Camera::zoom().
Camera::zoom() is an alternative to Camera::field_of_view() which is
linearized in distance (specifically the distance between the look at
and look from points), rather than the angle. Therefore, a value of 1.0
will place the screen at the look at point, while a value of 2.0 will
place the screen at the point twice as far as the look at point
(measured from the look from point). The input parameter is clamped from
below at 0.0.